Migrate sampling benchmark.
