+++ API/test_df.py
@@ -1,7 +1,8 @@
 import string
 
+import numpy
 import pytest
-from config import cudf
+from config import cudf, cupy
 
 exprs = ["a+b", "a+b+c+d+e", "a / (sin(a) + cos(b)) * tan(d*e*f)"]
 
@@ -36,3 +37,21 @@ def test_merge(benchmark, dataframe_dtype_int_cols_6, nkey_cols):
 )
 def test_isin(benchmark, dataframe_dtype_int, values):
     benchmark(dataframe_dtype_int.isin, values)
+
+
+@pytest.fixture(
+    params=[0, numpy.random.RandomState, cupy.random.RandomState],
+    ids=["Seed", "NumpyRandomState", "CupyRandomState"],
+)
+def random_state(request):
+    rs = request.param
+    return rs if isinstance(rs, int) else rs(seed=42)
+
+
+@pytest.mark.parametrize("frac", [0.5])
+def test_sample(benchmark, dataframe_dtype_int, axis, frac, random_state):
+    if axis == 1 and isinstance(random_state, cupy.random.RandomState):
+        pytest.skip("Unsupported params.")
+    benchmark(
+        dataframe_dtype_int.sample, frac=frac, axis=axis, random_state=random_state
+    )
